requirements-dev.txt
asgiref==3.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.5
djangorestframework==3.12.2
python-decouple==3.4
python-http-client==3.3.1
pytz==2020.5
sendgrid==6.4.8
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
static3==0.7.0
